+++ src/main/java/com/kusitms/wannafly/command/auth/infrastructure/refreshtoken/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.kusitms.wannafly.command.auth.infrastructure.refreshtoken;
+package com.kusitms.wannafly.command.auth.infrastructure.refreshtoken.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
