+++ src/main/java/com/kusitms/wannafly/command/auth/infrastructure/refreshtoken/redis/RefreshTokenRedisRepository.java
@@ -1,4 +1,4 @@
-package com.kusitms.wannafly.command.auth.infrastructure.refreshtoken;
+package com.kusitms.wannafly.command.auth.infrastructure.refreshtoken.redis;
 
 import org.springframework.data.repository.CrudRepository;
